refactor: move proto serialization into its own module
Both copies of the ChatItem mapping now sit in one file, which is the point:
they were written twice, in voice and in telemetry, and had drifted.

The move is otherwise verbatim, except that the voice copy dispatches on
ChatItem's discriminator rather than isinstance. Importing the classes would
pull llm in, and llm reaches telemetry, so the module could not stay a leaf.
